Reorganized Shopping List page's loading skeleton into a component and match the skeleton to what page contains
